Modifies mock_api for demo.
